Refactored and made sure right lstm goes backwards
