+++ bella/models/tdlstm.py
@@ -1,25 +1,13 @@
-import random as rn
-import os
-import pickle
-import tempfile
-import time
 from pathlib import Path
 from typing import Dict, Callable, Any, List, Union, Tuple
 
 import numpy as np
-import tensorflow as tf
-import keras
-from keras import backend as K
-from keras import preprocessing, models, optimizers, initializers, layers
-from keras.callbacks import EarlyStopping, ModelCheckpoint
+from keras import models, optimizers, layers
 from keras.utils import to_categorical
-from keras.models import model_from_yaml
 
 # Displaying the Neural Network models
 from keras.utils.vis_utils import model_to_dot, plot_model
 from IPython.display import SVG
-
-from sklearn.model_selection import StratifiedShuffleSplit, StratifiedKFold
 
 import bella
 from bella.models.base import KerasModel
@@ -117,8 +105,12 @@ class LSTM(KerasModel):
 
     def model_parameters(self) -> Dict[str, Any]:
         '''
-        Returns a dictionary of all of the attributes that affect the model as
-        well as the class the model belongs to.
+        Returns a dictionary containing the attributes of the class instance,
+        the parameters to give to the class constructior to re-create this
+        instance, and the class itself.
+
+        This is used by the :py:meth:`save` method so that the instance can
+        be re-created when loaded by the :py:meth:`load` method.
         '''
 
         class_params = {'tokeniser': self.tokeniser,
@@ -137,132 +129,69 @@ class LSTM(KerasModel):
         return {'class': self.__class__, 'class_attrs': class_attrs,
                 'class_params': class_params}
 
-    def process_text(self, texts, max_length, padding='pre', truncate='pre'):
+    def _pre_process(self, data_dicts: Dict[str, str],
+                     training: bool = False) -> List[List[int]]:
         '''
-        Given a list of Strings, tokenised the text and lower case if set and then
-        convert the tokens into a integers representing the tokens in the
-        embeddings. Lastly it pads the data based on the max_length param.
+        Converts the text in the data_dicts into a list of a list of integers
+        representing the text as their embedding lookups, so that it can
+        be used as input to the keras model.
 
-        If the max_length is smaller than the sentences size it truncates the
-        sentence. If max_length = -1 then the max_length is that of the longest
-        sentence in the texts.
+        The text from the data_dicts are converted by the
+        :py:meth:`process_text` method.
 
-        :params texts: list of Strings
-        :params max_length: How many tokens a sentence can contain. If it is \
-        -1 then it uses the sentence with the most tokens as the max_length param.
-        :params padding: which side of the sentence to pad: `pre` beginning, \
-        `post` end.
-        :params truncate: which side of the sentence to truncate: `pre` beginning \
-        `post` end.
-        :type texts: list
-        :type max_length: int
-        :type padding: String. Either `pre` or `post` default `pre`
-        :type truncate: String. Either `pre` or `post` default `pre`
-        :returns: A tuple of length 2 containg: 1. The max_length parameter, 2. \
-        A list of a list of integers that have been padded.
-        :rtype: tuple
+        :param data_dicts: A list of dictonaries that contain a `text` key.
+        :param training: Whether the text should be processed for training or
+                         for prediction. prediction = False, training = True
+        :return: The output of :py:meth:`process_text` method.
         '''
-
-        if max_length == 0:
-            raise ValueError('The max length of a sequence cannot be zero')
-        elif max_length < -1:
-            raise ValueError('The max length has to be either -1 or above '
-                             f'zero not {max_length}')
-
-        # Process the text into integers based on the embeddings given
-        all_sequence_data = []
-        max_sequence = 0
-        for text in texts:
-            sequence_data = []
-            tokens = self.tokeniser(text)
-            for token in tokens:
-                if self.lower:
-                    token = token.lower()
-                sequence_data.append(self.embeddings.word2index[token])
-            sequence_length = len(sequence_data)
-            if sequence_length > max_sequence:
-                max_sequence = sequence_length
-            all_sequence_data.append(sequence_data)
-        if max_sequence == 0:
-            raise ValueError('The max sequence length is 0 suggesting no '\
-                             'data was provided for training or testing')
-        # Pad the sequences
-        # If max pad size is set and training the model set the test_pad_size
-        # to max sequence length
-        if max_length == -1:
-            max_length = max_sequence
-        return (max_length,
-                preprocessing.sequence.pad_sequences(all_sequence_data,
-                                                     maxlen=max_length,
-                                                     dtype='int32',
-                                                     padding=padding,
-                                                     truncating=truncate))
-
-    def _pre_process(self, data_dicts, training=False):
         text_data = [data['text'] for data in data_dicts]
         if training:
             pad_data = self.process_text(text_data, self.pad_size)
             self.test_pad_size, sequence_data = pad_data
             return sequence_data
-        else:
-            _, sequence_data = self.process_text(text_data, self.test_pad_size)
-            return sequence_data
+        _, sequence_data = self.process_text(text_data, self.test_pad_size)
+        return sequence_data
 
-    def create_training_y(self, train_y, validation_y):
+    def create_training_y(self, train_y: np.ndarray, validation_y: np.ndarray,
+                          ) -> Tuple[np.ndarray, np.ndarray]:
+        '''
+        Converts the training and validation target values from a vector of
+        class lables into a matrix of binary values.
+
+        To convert the vector of classes to a matrix we the
+        :py:func:`keras.utils.to_categorical` function.
+
+        :param train_y: Vector of class labels, shape = [n_samples]
+        :param validation_y: Vector of class labels, shape = [n_samples]
+        :return: A tuple of length two containing the train and validation
+                 matrices respectively. The shape of each matrix is:
+                 [n_samples, n_classes]
+        '''
         train_y = to_categorical(train_y).astype(np.float32)
         validation_y = to_categorical(validation_y).astype(np.float32)
         return train_y, validation_y
 
-    def create_training_text(self, train_data, validation_data):
+    def create_training_text(self, train_data: Dict[str, str],
+                             validation_data: Dict[str, str]
+                             ) -> Tuple[List[List[int]], List[List[int]]]:
         '''
-        :param train_data: Training features. Specifically a list of dict like \
-        structures that contain `text` key.
-        :param train_y: Target values
-        :validation_size: The fraction of the training data to be set aside \
-        for validation data
-        :type train_data: list
-        :type train_y: list
-        :type validation_size: float Default 0.2
-        :returns: A tuple of length 2 where the first value is a list of \
-        Integers that reprsent the words in the text features where each Integer \
-        corresponds to a Word Vector in the embedding vector. Second value are \
-        the target values. Both lists in the tuples contain training data in the \
-        first part of the list and the second part of the list based on the \
-        validation split contains the validation data.
-        :rtype: tuple
+        Converts the training and validation data into a format that the keras
+        model can take as input.
+
+        :return: A tuple of length two containing the train and validation
+                 input that are both the output of :py:meth:`_pre_process`
         '''
+        train_sequence = self._pre_process(train_data, training=True)
+        val_sequence = self._pre_process(validation_data, training=False)
+        return train_sequence, val_sequence
 
-        # Convert from a sequence of dictionaries into texts and then integers
-        # that represent the tokens in the text within the embedding space.
-        sequence_train_data = self._pre_process(train_data, training=True)
-        sequence_val_data = self._pre_process(validation_data, training=False)
-        # Stack the validation data with the training data to complie with Keras.
-        # all_text = np.vstack((sequence_train_data, sequence_val_data))
-        return sequence_train_data, sequence_val_data
+    def keras_model(self, num_classes: int) -> 'keras.models.Model':
+        '''
+        The model that represents this class. This is a Single forward LSTM.
 
-    def repeated_results(self, train, test, n_results, score_func, dataset_name,
-                         score_args=None, score_kwargs=None,
-                         results_file=None, re_write=False, **fit_kwargs):
-        if results_file is not None:
-            all_scores = get_json_data(results_file, dataset_name)
-            if len(all_scores) != 0  and not re_write:
-                return all_scores
-        train_data = train.data_dict()
-        train_y = train.sentiment_data()
-        test_data = test.data_dict()
-        test_y = test.sentiment_data()
-        scores = []
-        for i in range(n_results):
-            score, _ = self.fit_predict(train_data, train_y, test_data,
-                                        test_y, fit_kwargs, score_func,
-                                        score_args, score_kwargs)
-            print(score)
-            scores.append(score)
-            if results_file is not None:
-                write_json_data(results_file, dataset_name, scores)
-        return scores
-
-    def keras_model(self, num_classes):
+        :param num_classes: Number of classes to predict.
+        :return: Forward LSTM keras model.
+        '''
         # Embeddings
         embedding_matrix = self.embeddings.embedding_matrix
         vocab_size, vector_size = embedding_matrix.shape
@@ -301,69 +230,6 @@ class LSTM(KerasModel):
         return models.Model(inputs=input_layer, outputs=prediction_layer)
 
     @property
-    def tokeniser(self) -> Callable[[str], List[str]]:
-        '''
-        tokeniser attribute
-
-        :return: The tokeniser used in the model
-        '''
-
-        return self._tokeniser
-
-    @tokeniser.setter
-    def tokeniser(self, value: Callable[[str], List[str]]) -> None:
-        '''
-        Sets the tokeniser attribute
-
-        :param value: The value to assign to the tokeniser attribute
-        '''
-
-        self.fitted = False
-        self._tokeniser = value
-
-    @property
-    def embeddings(self) -> 'bella.word_vectors.WordVectors':
-        '''
-        embeddings attribute
-
-        :return: The embeddings used in the model
-        '''
-
-        return self._embeddings
-
-    @embeddings.setter
-    def embeddings(self, value: 'bella.word_vectors.WordVectors') -> None:
-        '''
-        Sets the embeddings attribute
-
-        :param value: The value to assign to the embeddings attribute
-        '''
-
-        self.fitted = False
-        self._embeddings = value
-
-    @property
-    def reproducible(self) -> Union[int, None]:
-        '''
-        reproducible attribute
-
-        :return: The reproducible used in the model
-        '''
-
-        return self._reproducible
-
-    @reproducible.setter
-    def reproducible(self, value: Union[int, None]) -> None:
-        '''
-        Sets the reproducible attribute
-
-        :param value: The value to assign to the reproducible attribute
-        '''
-
-        self.fitted = False
-        self._reproducible = value
-
-    @property
     def pad_size(self) -> int:
         '''
         pad_size attribute
@@ -383,90 +249,6 @@ class LSTM(KerasModel):
 
         self.fitted = False
         self._pad_size = value
-
-    @property
-    def lower(self) -> bool:
-        '''
-        lower attribute
-
-        :return: The lower used in the model
-        '''
-
-        return self._lower
-
-    @lower.setter
-    def lower(self, value: bool) -> None:
-        '''
-        Sets the lower attribute
-
-        :param value: The value to assign to the lower attribute
-        '''
-
-        self.fitted = False
-        self._lower = value
-
-    @property
-    def patience(self) -> int:
-        '''
-        patience attribute
-
-        :return: The patience used in the model
-        '''
-
-        return self._patience
-
-    @patience.setter
-    def patience(self, value: int) -> None:
-        '''
-        Sets the patience attribute
-
-        :param value: The value to assign to the patience attribute
-        '''
-
-        self.fitted = False
-        self._patience = value
-
-    @property
-    def batch_size(self) -> int:
-        '''
-        batch_size attribute
-
-        :return: The batch_size used in the model
-        '''
-
-        return self._batch_size
-
-    @batch_size.setter
-    def batch_size(self, value: int) -> None:
-        '''
-        Sets the batch_size attribute
-
-        :param value: The value to assign to the batch_size attribute
-        '''
-
-        self.fitted = False
-        self._batch_size = value
-
-    @property
-    def epochs(self) -> int:
-        '''
-        epochs attribute
-
-        :return: The epochs used in the model
-        '''
-
-        return self._epochs
-
-    @epochs.setter
-    def epochs(self, value: int) -> None:
-        '''
-        Sets the epochs attribute
-
-        :param value: The value to assign to the epochs attribute
-        '''
-
-        self.fitted = False
-        self._epochs = value
 
     @property
     def embedding_layer_kwargs(self) -> Dict[str, Any]:
@@ -533,29 +315,6 @@ class LSTM(KerasModel):
 
         self.fitted = False
         self._dense_layer_kwargs = value
-
-    @property
-    def optimiser(self) -> 'keras.optimizers.Optimizer':
-        '''
-        optimiser attribute
-
-        :return: The optimiser used in the model
-        '''
-
-        return self._optimiser
-
-    @optimiser.setter
-    def optimiser(self, value: 'keras.optimizers.Optimizer') -> None:
-        '''
-        Sets the optimiser attribute
-
-        :param value: The value to assign to the optimiser attribute
-        '''
-
-        self.fitted = False
-        self._optimiser = value
-
-
 
     def visulaise(self, plot_format='vert'):
         '''
@@ -673,8 +432,12 @@ class TDLSTM(LSTM):
 
     def model_parameters(self) -> Dict[str, Any]:
         '''
-        Returns a dictionary of all of the attributes that affect the model as
-        well as the class the model belongs to.
+        Returns a dictionary containing the attributes of the class instance,
+        the parameters to give to the class constructior to re-create this
+        instance, and the class itself.
+
+        This is used by the :py:meth:`save` method so that the instance can
+        be re-created when loaded by the :py:meth:`load` method.
         '''
 
         attributes = super().model_parameters()
@@ -720,17 +483,17 @@ class TDLSTM(LSTM):
                                                    truncate='post')
             self.right_test_pad_size, right_sequence = right_pad_sequence
             return left_sequence, right_sequence
-        else:
-            left_pad_sequence = self.process_text(left_text,
-                                                  self.left_test_pad_size)
-            _, left_sequence = left_pad_sequence
 
-            right_pad_sequence = self.process_text(right_text,
-                                                   self.right_test_pad_size,
-                                                   padding='post',
-                                                   truncate='post')
-            _, right_sequence = right_pad_sequence
-            return [left_sequence, right_sequence]
+        left_pad_sequence = self.process_text(left_text,
+                                              self.left_test_pad_size)
+        _, left_sequence = left_pad_sequence
+
+        right_pad_sequence = self.process_text(right_text,
+                                               self.right_test_pad_size,
+                                               padding='post',
+                                               truncate='post')
+        _, right_sequence = right_pad_sequence
+        return [left_sequence, right_sequence]
 
     def create_training_text(self, train_data, validation_data):
         '''
@@ -751,12 +514,18 @@ class TDLSTM(LSTM):
         :rtype: tuple
         '''
 
-        train_sequences = self._pre_process(train_data, training=True)
-        validation_sequences = self._pre_process(validation_data,
-                                                 training=False)
-        return train_sequences, validation_sequences
+        return super().create_training_text(train_data, validation_data)
 
-    def keras_model(self, num_classes):
+    def keras_model(self, num_classes: int) -> 'keras.models.Model':
+        '''
+        The model that represents this class. This is a custom combination
+        of two LSTMs.
+
+        :param num_classes: Number of classes to predict.
+        :return: Two LSTMs, one forward from the left context and the other
+                 backward from the right context. The output of the two are
+                 concatenated and are input to the output layer.
+        '''
         # Embeddings
         embedding_matrix = self.embeddings.embedding_matrix
         vocab_size, vector_size = embedding_matrix.shape
@@ -803,6 +572,7 @@ class TDLSTM(LSTM):
                                            )(right_input)
         right_lstm_layer = layers.LSTM(lstm_dimension,
                                        name='right_lstm_layer',
+                                       go_backwards=True,
                                        **lstm_layer_kwargs
                                        )(right_embedding_layer)
         # Merge the outputs of the left and right LSTMs
@@ -848,62 +618,6 @@ class TCLSTM(TDLSTM):
 
         return 'TCLSTM'
 
-    def __init__(self, tokeniser: Callable[[str], List[str]],
-                 embeddings: 'bella.word_vectors.WordVectors',
-                 reproducible: Union[int, None] = None, pad_size: int = -1,
-                 lower: bool = True, patience: int = 10,
-                 batch_size: int = 32, epochs: int = 300,
-                 embedding_layer_kwargs: Dict[str, Any] = None,
-                 lstm_layer_kwargs: Dict[str, Any] = None,
-                 dense_layer_kwargs: Dict[str, Any] = None,
-                 optimiser: 'keras.optimizers.Optimizer' = optimizers.SGD,
-                 optimiser_params: Union[Dict[str, Any], None] = None,
-                 include_target: bool = True) -> None:
-        '''
-        :param tokeniser: Tokeniser to be used e.g. :py:meth:`str.split`.
-        :param embeddings: Embedding (Word vectors) to be used e.g.
-                           :py:class:`bella.word_vectors.SSWE`
-        :param reproducible: Whether to be reproducible. If None then it is
-                             but quicker to run. Else provide a `int` that
-                             will represent the random seed value.
-        :param pad_size: The max number of tokens to use per sequence. If -1
-                         use the text sequence in the training data that has
-                         the most tokens as the pad size.
-        :param lower: Whether to lower case the words being processed.
-        :param patience: Number of epochs with no improvement before training
-                         is stopped.
-        :param batch_size: Number of samples per gradient update.
-        :param epochs: Number of times to train over the entire training set
-                       before stopping. If patience is set, then it may
-                       stop before reaching the number of epochs specified
-                       here.
-        :param embedding_layer_kwargs: Keyword arguments to pass to the
-                                       embedding layer which is a
-                                       :py:class:`keras.layers.Embedding`
-                                       object. If no parameters to pass leave
-                                       as None.
-        :param lstm_layer_kwargs: Keyword arguments to pass to the lstm
-                                  layer(s) which is a
-                                  :py:class:`keras.layers.LSTM` object. If no
-                                  parameters to pass leave as None.
-        :param dense_layer_kwargs: Keyword arguments to pass to the dense
-                                   (final layer) which is a
-                                   :py:class:`keras.layers.Dense` object. If no
-                                   parameters to pass leave as None.
-        :param optimiser: Optimiser to be used accepts any
-                          `keras optimiser <https://keras.io/optimizers/>`_.
-                          Default is :py:class:`keras.optimizers.SGD`
-        :param optimiser_params: Parameters for the optimiser. If None uses
-                                 default optimiser parameters.
-        :param include_target: Wheather to include the target in the LSTM
-                               representations.
-        '''
-
-        super().__init__(tokeniser, embeddings, reproducible, pad_size, lower,
-                         patience, batch_size, epochs, embedding_layer_kwargs,
-                         lstm_layer_kwargs, dense_layer_kwargs, optimiser,
-                         optimiser_params, include_target)
-
     def _pre_process(self, data_dicts, training=False):
         def context_median_targets(pad_size):
             vector_size = self.embeddings.vector_size
@@ -938,10 +652,7 @@ class TCLSTM(TDLSTM):
         a list of dict like structures that contain `target` key.
         '''
 
-        train_seq_targ = self._pre_process(train_data, training=True)
-        validation_seq_targ = self._pre_process(validation_data,
-                                                training=False)
-        return train_seq_targ, validation_seq_targ
+        return super().create_training_text(train_data, validation_data)
 
     def keras_model(self, num_classes):
         # Embeddings
@@ -1004,6 +715,7 @@ class TCLSTM(TDLSTM):
                                                name='right_text_target')
         right_lstm_layer = layers.LSTM(lstm_dimension,
                                        name='right_lstm_layer',
+                                       go_backwards=True,
                                        **lstm_layer_kwargs
                                        )(right_text_target)
         # Merge the outputs of the left and right LSTMs
